packages/extension/pyproject.toml
[tool.poetry]
name = "lnbits-extension"
version = "0.11.9"
description = "LNbits, free and open-source Lightning wallet and accounts system."
authors = ["Alan Bits <alan@lnbits.com>"]
packages = [
  { include = "lnbits_extension" },
]

[tool.poetry.dependencies]
python = "^3.10 | ^3.9"
bech32 = "1.2.0"
bitstring = "3.1.9"
click = "8.0.4"
ecdsa = "0.18.0"
embit = "0.4.9"
fastapi = "^0"
httpx = "0.23.0"
jinja2 = "3.0.1"
lnurl = "0.3.6"
psycopg2-binary = "2.9.1"
pydantic = "1.10.4"
pyqrcode = "1.2.1"
shortuuid = "1.0.11"
sqlalchemy = "1.3.24"
sqlalchemy-aio = "0.17.0"
sse-starlette = "0.6.2"
typing-extensions = "4.4.0"
uvicorn = "0.18.3"
uvloop = "0.16.0"
websockets = "10.0"
loguru = "0.6.0"
grpcio = "1.51.1"
protobuf = "4.21.12"
Cerberus = "1.3.4"
async-timeout = "4.0.2"
pyln-client = "0.11.1"
cashu = "0.9.0"
docker = "^6.1.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.1.2"
mock = "^4.0.3"
black = "^22.6.0"
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"
mypy = "^0.971"
types-protobuf = "^3.19.22"
pylint = "^2.17.2"
pre-commit = "^3.2.2"
types-mock = "^5.0.0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
lnbits = "lnbits.server:main"

[tool.isort]
profile = "black"

[tool.pyright]
include = [
  "lnbits"
]
exclude = [
  "lnbits/wallets/lnd_grpc_files",
  "lnbits/extensions",
]

[tool.mypy]
files = "lnbits"
exclude = [
  "^lnbits/wallets/lnd_grpc_files",
  "^lnbits/extensions",
]

[[tool.mypy.overrides]]
module = [
  "embit.*",
  "secp256k1.*",
  "uvicorn.*",
  "sqlalchemy.*",
  "sqlalchemy_aio.*",
  "websocket.*",
  "websockets.*",
  "pyqrcode.*",
  "shortuuid.*",
  "grpc.*",
  "lnurl.*",
  "bitstring.*",
  "ecdsa.*",
  "psycopg2.*",
  "pyngrok.*",
]
ignore_missing_imports = "True"

[tool.pytest.ini_options]
addopts = "--durations=1 -s --cov=lnbits --cov-report=xml"
testpaths = [
  "tests"
]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 300
disable = "all"
enable = [
  "assignment-from-none",
  "chained-comparison",
  "consider-merging-isinstance",
  "consider-using-dict-comprehension",
  "consider-using-dict-items",
  "consider-using-f-string",
  "consider-using-in",
  "dangerous-default-value",
  "inconsistent-return-statements",
  "lost-exception",
  "pointless-string-statement",
  "simplifiable-if-statement",
  "super-init-not-called",
  "superfluous-parens",
  "unused-variable",
  "use-list-literal",
  "useless-else-on-loop",
  "useless-object-inheritance",
]

[tool.pylint.MASTER]
ignore-paths = [
  "^lnbits/extensions/.*$",
  "^lnbits/wallets/lnd_grpc_files/.*$",
]
fail-under = 10.0
jobs = 0
